third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-sizing/contain-intrinsic-size/auto-007-expected.txt

This is a testharness.js-based test.
PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
FAIL .test 5 assert_equals: 
<div class="scroll test auto-height" data-expected-client-width="0" data-expected-client-height="255"></div>
clientHeight expected 255 but got 270
FAIL .test 6 assert_equals: 
<div class="scroll test auto-both" data-expected-client-width="320" data-expected-client-height="255"></div>
clientHeight expected 255 but got 270
FAIL .test 7 assert_equals: 
<div class="columns test auto-width" data-expected-client-width="656" data-expected-client-height="0"></div>
clientWidth expected 656 but got 1328
PASS .test 8
FAIL .test 9 assert_equals: 
<div class="columns test auto-both" data-expected-client-width="656" data-expected-client-height="120"></div>
clientWidth expected 656 but got 1328
PASS .test 10
PASS .test 11
PASS .test 12
PASS .test 13
PASS .test 14
PASS .test 15
PASS .test 16
PASS .test 17
PASS .test 18
PASS .test 19
PASS .test 20
PASS .test 21
PASS .test 22
PASS .test 23
PASS .test 24
FAIL .test 25 assert_equals: 
<canvas class="test auto-width" data-expected-client-width="300" data-expected-client-height="0"></canvas>
clientWidth expected 300 but got 1
FAIL .test 26 assert_equals: 
<canvas class="test auto-height" data-expected-client-width="0" data-expected-client-height="150"></canvas>
clientHeight expected 150 but got 2
FAIL .test 27 assert_equals: 
<canvas class="test auto-both" data-expected-client-width="300" data-expected-client-height="150"></canvas>
clientWidth expected 300 but got 3
PASS .test 28
PASS .test 29
PASS .test 30
PASS .test 31
PASS .test 32
PASS .test 33
PASS .test 34
PASS .test 35
PASS .test 36
PASS .test 37
PASS .test 38
PASS .test 39
PASS .test 40
PASS .test 41
PASS .test 42
Harness: the test ran to completion.
